+++ eleventy.config.js
@@ -186,6 +186,7 @@ export default async function (eleventyConfig) {
     "src/assets/js": "assets/js",
     "node_modules/alpinejs/dist/cdn.min.js": "assets/js/alpine.min.js",
     "node_modules/@tailwindplus/elements/dist/index.js": "assets/js/elements.js",
+    "content/guidelines/design.md": "design.md",
     "content/guidelines": "guidelines",
   });